+++ FO-DICOM.Core/Serialization/JsonDicomConverter.cs
@@ -612,9 +612,17 @@ namespace FellowOakDicom.Serialization
                     {
                         writer.WriteNullValue();
                     }
-                    else if (val is float f && float.IsNaN(f))
+                    else if ((val is float f && float.IsNaN(f)) || (val is double d && double.IsNaN(d)))
                     {
                         writer.WriteStringValue("NaN");
+                    }
+                    else if ((val is double dp && double.IsPositiveInfinity(dp)) || (val is float fp && float.IsPositiveInfinity(fp)))
+                    {
+                        writer.WriteStringValue("Infinity");
+                    }
+                    else if ((val is double dn && double.IsNegativeInfinity(dn)) || (val is float fn && float.IsNegativeInfinity(fn)))
+                    {
+                        writer.WriteStringValue("-Infinity");
                     }
                     else
                     {
@@ -901,7 +909,36 @@ namespace FellowOakDicom.Serialization
                 }
                 else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "NaN")
                 {
-                    childValues.Add((T)(float.NaN as object));
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.NaN as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.NaN as object));
+                    }
+                }
+                else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "Infinity")
+                {
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.PositiveInfinity as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.PositiveInfinity as object));
+                    }
+                }
+                else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "-Infinity")
+                {
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.NegativeInfinity as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.NegativeInfinity as object));
+                    }
                 }
                 else if (reader.TokenType == JsonTokenType.String && tryParse(reader.GetString(), out T parsed))
                 {
@@ -956,7 +993,36 @@ namespace FellowOakDicom.Serialization
                 }
                 else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "NaN")
                 {
-                    childValues.Add((T)(float.NaN as object));
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.NaN as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.NaN as object));
+                    }
+                }
+                else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "Infinity")
+                {
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.PositiveInfinity as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.PositiveInfinity as object));
+                    }
+                }
+                else if (reader.TokenType == JsonTokenType.String && reader.GetString() == "-Infinity")
+                {
+                    if (typeof(T) == typeof(double))
+                    {
+                        childValues.Add((T)(double.NegativeInfinity as object));
+                    }
+                    else
+                    {
+                        childValues.Add((T)(float.NegativeInfinity as object));
+                    }
                 }
                 else
                 {