+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 - Added core support for HTJ2K-based transfer syntaxes (not actual codec) (#1687)
 - Add support for parsing DICOM files where the pixel data is not properly closed with a SequenceDelimitationItem (#1339)
+- Update Dicom json converter to handle Infinity values for FL and FD VRs (#1725)
 
 ### 5.1.2 (2023-12-21)